implement code splitting, lazy loading and SVG in place of png image formats
