+++ package.json
@@ -6,16 +6,12 @@
     "@testing-library/jest-dom": "^5.16.1",
     "@testing-library/react": "^12.1.2",
     "@testing-library/user-event": "^13.5.0",
-    "bootstrap": "^5.1.3",
     "daisyui": "^2.11.0",
     "emulators-ui": "^0.73.1",
     "js-dos": "^7.3.7",
     "react": "^17.0.2",
-    "react-bootstrap": "^2.0.3",
     "react-dom": "^17.0.2",
     "react-draggable": "^4.4.4",
-    "react-modal-resizable-draggable": "^0.1.6",
-    "react-rnd": "^10.3.5",
     "react-router-dom": "^6.2.1",
     "react-scripts": "5.0.0",
     "sass": "^1.45.0",
