+++ public/index.html
@@ -8,19 +8,19 @@
 		<meta name="description" content="A portfolio website to showcase my work" />
 		<link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
 
-		<script src="%PUBLIC_URL%/js-dos/js-dos.js"></script>
 		<link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
 
 		<!-- JS DOS -->
 
 		<link rel="stylesheet" href="%PUBLIC_URL%/js-dos/js-dos.css" />
-		<script>
-			emulators.pathPrefix = "%PUBLIC_URL%/js-dos/";
-		</script>
 		<title>Karl Swatman | Full Stack Developer</title>
 	</head>
 	<body>
 		<noscript>You need to enable JavaScript to run this app.</noscript>
 		<div id="root"></div>
 	</body>
+	<script src="%PUBLIC_URL%/js-dos/js-dos.js"></script>
+	<script>
+		emulators.pathPrefix = "%PUBLIC_URL%/js-dos/";
+	</script>
 </html>
