+++ src/App.js
@@ -1,10 +1,7 @@
-import { useRef, useEffect, useState } from "react";
+import { useRef } from "react";
 import "./App.css";
 import { Route, Routes } from "react-router-dom";
-import { Header } from "./components/Header";
 import { NavBar } from "./components/NavBar";
-import { WorkCard } from "./components/WorkCard";
-import { Home } from "./pages/Home";
 import Desktop from "./pages/Desktop";
 
 function App() {
@@ -17,7 +14,6 @@ function App() {
 		>
 			<NavBar />
 			<Routes>
-				<Route path="/" element={<Home />} />
 				<Route path="/desktop" element={<Desktop bg={backgroundRef} />} />
 			</Routes>
 		</div>
