+++ src/components/Icon.jsx
@@ -1,6 +1,4 @@
 import React from "react";
-import ReactDOM from "react-dom";
-import { Header } from "../components/Header";
 
 const Icon = ({
 	children,
