+++ src/components/NavBar.jsx
@@ -1,4 +1,4 @@
-import logo from "../assets/images/logo.png";
+import logo from "../assets/images/logo.svg";
 
 export const NavBar = () => {
 	return (
