--- src/pages/Home.jsx
@@ -1,11 +1,0 @@
-import { Header } from "../components/Header";
-import { WorkCard } from "../components/WorkCard";
-
-export const Home = () => {
-	return (
-		<>
-			<Header />
-			<WorkCard />
-		</>
-	);
-};
